+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add search bar, refresh button, and pagination state persistence, [PR-135](https://github.com/reductstore/web-console/pull/135)
 - Added expandable table rows to display and edit labels directly within the table view, [PR-136](https://github.com/reductstore/web-console/pull/136)
 - Use JSON5 to set `when` conditions, [PR-139](https://github.com/reductstore/web-console/pull/139)
+- Added record preview for text and images in the entry detail view, [PR-141](https://github.com/reductstore/web-console/pull/141)
 
 ### Fixed